+++ include/FileParser/Jpeg/JpegBitWriter.h
@@ -3,7 +3,7 @@
 
 class JpegBitWriter final : public BitWriter{
 public:
-    explicit JpegBitWriter(const std::string& filepath, int bufferSize = 4096) : BitWriter(filepath, bufferSize) {}
+    explicit JpegBitWriter(const std::string& filepath, size_t bufferSize = 4096) : BitWriter(filepath, bufferSize) {}
     
     void setByteStuffing(const bool byteStuffing) {
         m_byteStuffing = byteStuffing;
